use reportProgress tool in a more robust way (#39)
*Issue #, if available:*

*Description of changes:*

This PR tries to resolve the problem that an agent does not send any
message for a while after user's message, making users confused if the
agent is stuck or not.

To improve the situation, we use
[toolConfig](https://docs.aws.amazon.com/bedrock/latest/APIReference/API_runtime_ToolConfiguration.html#bedrock-Type-runtime_ToolConfiguration-tools)
to force the agent to use reportProgress tool. To let the agent decide
not to use the tool, we also added a think tool (kind of no-op tool in
this case, but it will also improve the agents' thought capability), and
toolConfig any option.

Because toolConfig option cannot be enabled with reasoning option, we
need special handling for the props.



By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.