Monorepo (#48)
*Issue #, if available:*

*Description of changes:*

Switch to monorepo structure to allow to share code between modules.
Currently shared code is extracted to `agent-core` package, which can be
referenced from `slack-bolt-app` and `worker`. `cdk` does not require to
be itself in monorepo, so it is left isolated now. We might move cdk
under monorepo in the future.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.